# admin-web/src/app/core/api/generated/model/createRoleRequest.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Pointer.API
 * OpenAPI spec version: 1.0
 */

export interface CreateRoleRequest {
  /** @nullable */
  name?: string | null;
  grantsAdmin?: boolean;
}