# admin-web/src/app/core/api/generated/model/projectResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Pointer.API
 * OpenAPI spec version: 1.0
 */

export interface ProjectResponse {
  id?: number;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  name?: string | null;
  isActive?: boolean;
}