# admin-web/src/app/core/api/generated/model/roleResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Pointer.API
 * OpenAPI spec version: 1.0
 */

export interface RoleResponse {
  id?: number;
  /** @nullable */
  name?: string | null;
  grantsAdmin?: boolean;
  isSystem?: boolean;
  isActive?: boolean;
}