# admin-web/src/app/core/api/generated/model/updateUserRequest.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Pointer.API
 * OpenAPI spec version: 1.0
 */

export interface UpdateUserRequest {
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  password?: string | null;
}