# admin-web/src/app/core/api/generated/model/userResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Pointer.API
 * OpenAPI spec version: 1.0
 */
import type { ApprovalStatus } from './approvalStatus';

export interface UserResponse {
  id?: number;
  publicId?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  displayName?: string | null;
  roleId?: number;
  /** @nullable */
  roleName?: string | null;
  isAdmin?: boolean;
  isActive?: boolean;
  approvalStatus?: ApprovalStatus;
}